refactor(maestro): split server/state into cohesive modules
