--- crates/vize_maestro/src/server/state.rs
@@ -1,1574 +1,0 @@
-//! Server state management.
-#![allow(clippy::disallowed_types, clippy::disallowed_methods)]
-
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-use std::sync::atomic::{AtomicBool, Ordering};
-
-use dashmap::DashMap;
-use futures::lock::Mutex as AsyncMutex;
-use parking_lot::RwLock;
-use serde::Deserialize;
-use tower_lsp::lsp_types::Url;
-use vize_carton::config::{LanguageServerConfig, LinterConfig, TypeCheckerConfig};
-
-#[cfg(feature = "glyph")]
-use vize_carton::config::FormatterConfig;
-
-#[cfg(feature = "native")]
-use std::sync::OnceLock;
-
-#[cfg(feature = "native")]
-use vize_canon::{
-    BatchTypeChecker, BatchTypeCheckerOptions, BatchTypeCheckerTrait, CorsaBridge,
-    CorsaBridgeConfig,
-};
-
-use crate::document::DocumentStore;
-use crate::utils::is_standalone_html_path;
-use crate::virtual_code::{VirtualCodeGenerator, VirtualDocuments};
-
-#[derive(Debug, Default, Deserialize)]
-#[serde(default, rename_all = "camelCase")]
-struct LspConfigSection {
-    enabled: Option<bool>,
-    /// Legacy diagnostics switch. Kept as a lint diagnostics alias for older configs.
-    diagnostics: Option<bool>,
-    lint: Option<bool>,
-    typecheck: Option<bool>,
-    editor: Option<bool>,
-    ecosystem: Option<bool>,
-    options_api: Option<bool>,
-    legacy_vue2: Option<bool>,
-    completion: Option<bool>,
-    hover: Option<bool>,
-    definition: Option<bool>,
-    references: Option<bool>,
-    document_symbols: Option<bool>,
-    workspace_symbols: Option<bool>,
-    code_actions: Option<bool>,
-    rename: Option<bool>,
-    formatting: Option<bool>,
-    code_lens: Option<bool>,
-    semantic_tokens: Option<bool>,
-    document_links: Option<bool>,
-    folding_ranges: Option<bool>,
-    inlay_hints: Option<bool>,
-    file_rename: Option<bool>,
-    corsa: Option<bool>,
-    tsgo: Option<bool>,
-    cross_file: Option<bool>,
-}
-
-impl LspConfigSection {
-    fn apply_to(self, features: &mut LspFeatureConfig) {
-        if self.enabled == Some(false) {
-            *features = LspFeatureConfig::disabled();
-            return;
-        }
-
-        if let Some(enabled) = self.diagnostics {
-            features.lint = enabled;
-            if !enabled {
-                features.typecheck = false;
-            }
-        }
-
-        if let Some(enabled) = self.lint {
-            features.lint = enabled;
-        }
-
-        if let Some(enabled) = self.typecheck {
-            features.typecheck = enabled;
-        }
-        if self.corsa == Some(true) || self.tsgo == Some(true) {
-            features.typecheck = true;
-        }
-
-        if let Some(enabled) = self.editor {
-            features.apply_editor_bundle(enabled);
-        }
-
-        if let Some(enabled) = self.ecosystem {
-            features.ecosystem = enabled;
-        }
-
-        if let Some(enabled) = self.options_api {
-            features.options_api = enabled;
-        }
-
-        if let Some(enabled) = self.legacy_vue2 {
-            features.legacy_vue2 = enabled;
-        }
-
-        if let Some(enabled) = self.completion {
-            features.completion = enabled;
-        }
-        if let Some(enabled) = self.hover {
-            features.hover = enabled;
-        }
-        if let Some(enabled) = self.definition {
-            features.definition = enabled;
-        }
-        if let Some(enabled) = self.references {
-            features.references = enabled;
-        }
-        if let Some(enabled) = self.document_symbols {
-            features.document_symbols = enabled;
-        }
-        if let Some(enabled) = self.workspace_symbols {
-            features.workspace_symbols = enabled;
-        }
-        if let Some(enabled) = self.code_actions {
-            features.code_actions = enabled;
-        }
-        if let Some(enabled) = self.rename {
-            features.rename = enabled;
-        }
-        if let Some(enabled) = self.formatting {
-            features.formatting = enabled;
-        }
-        if let Some(enabled) = self.code_lens {
-            features.code_lens = enabled;
-        }
-        if let Some(enabled) = self.semantic_tokens {
-            features.semantic_tokens = enabled;
-        }
-        if let Some(enabled) = self.document_links {
-            features.document_links = enabled;
-        }
-        if let Some(enabled) = self.folding_ranges {
-            features.folding_ranges = enabled;
-        }
-        if let Some(enabled) = self.inlay_hints {
-            features.inlay_hints = enabled;
-        }
-        if let Some(enabled) = self.file_rename {
-            features.file_rename = enabled;
-        }
-        if let Some(enabled) = self.cross_file {
-            features.cross_file = enabled;
-        }
-    }
-}
-
-impl From<LanguageServerConfig> for LspConfigSection {
-    fn from(config: LanguageServerConfig) -> Self {
-        Self {
-            enabled: config.enabled,
-            diagnostics: config.diagnostics,
-            lint: config.lint,
-            typecheck: config.typecheck,
-            editor: config.editor,
-            ecosystem: config.ecosystem,
-            options_api: None,
-            legacy_vue2: None,
-            completion: config.completion,
-            hover: config.hover,
-            definition: config.definition,
-            references: config.references,
-            document_symbols: config.document_symbols,
-            workspace_symbols: config.workspace_symbols,
-            code_actions: config.code_actions,
-            rename: config.rename,
-            formatting: config.formatting,
-            code_lens: config.code_lens,
-            semantic_tokens: config.semantic_tokens,
-            document_links: config.document_links,
-            folding_ranges: config.folding_ranges,
-            inlay_hints: config.inlay_hints,
-            file_rename: config.file_rename,
-            corsa: config.corsa,
-            tsgo: config.tsgo,
-            cross_file: config.cross_file,
-        }
-    }
-}
-
-/// Feature switches for Maestro LSP capabilities.
-///
-/// Non-opinionated diagnostics and editor features default to on. Formatting stays
-/// opt-in because it encodes style preferences and can overlap with project formatters.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct LspFeatureConfig {
-    pub(crate) lint: bool,
-    pub(crate) typecheck: bool,
-    pub(crate) ecosystem: bool,
-    pub(crate) options_api: bool,
-    pub(crate) legacy_vue2: bool,
-    pub(crate) completion: bool,
-    pub(crate) hover: bool,
-    pub(crate) definition: bool,
-    pub(crate) references: bool,
-    pub(crate) document_symbols: bool,
-    pub(crate) workspace_symbols: bool,
-    pub(crate) code_actions: bool,
-    pub(crate) rename: bool,
-    pub(crate) formatting: bool,
-    pub(crate) code_lens: bool,
-    pub(crate) semantic_tokens: bool,
-    pub(crate) document_links: bool,
-    pub(crate) folding_ranges: bool,
-    pub(crate) inlay_hints: bool,
-    pub(crate) file_rename: bool,
-    pub(crate) cross_file: bool,
-}
-
-impl LspFeatureConfig {
-    fn disabled() -> Self {
-        Self {
-            lint: false,
-            typecheck: false,
-            ecosystem: false,
-            options_api: false,
-            legacy_vue2: false,
-            completion: false,
-            hover: false,
-            definition: false,
-            references: false,
-            document_symbols: false,
-            workspace_symbols: false,
-            code_actions: false,
-            rename: false,
-            formatting: false,
-            code_lens: false,
-            semantic_tokens: false,
-            document_links: false,
-            folding_ranges: false,
-            inlay_hints: false,
-            file_rename: false,
-            cross_file: false,
-        }
-    }
-
-    pub(crate) fn has_diagnostics(self) -> bool {
-        self.lint || self.typecheck || self.ecosystem
-    }
-
-    fn apply_editor_bundle(&mut self, enabled: bool) {
-        self.ecosystem = enabled;
-        self.completion = enabled;
-        self.hover = enabled;
-        self.definition = enabled;
-        self.references = enabled;
-        self.document_symbols = enabled;
-        self.workspace_symbols = enabled;
-        self.rename = enabled;
-        self.code_lens = enabled;
-        self.semantic_tokens = enabled;
-        self.document_links = enabled;
-        self.folding_ranges = enabled;
-        self.inlay_hints = enabled;
-        self.file_rename = enabled;
-    }
-}
-
-impl Default for LspFeatureConfig {
-    fn default() -> Self {
-        Self {
-            lint: true,
-            typecheck: true,
-            ecosystem: true,
-            options_api: false,
-            legacy_vue2: false,
-            completion: true,
-            hover: true,
-            definition: true,
-            references: true,
-            document_symbols: true,
-            workspace_symbols: true,
-            code_actions: true,
-            rename: true,
-            formatting: false,
-            code_lens: true,
-            semantic_tokens: true,
-            document_links: true,
-            folding_ranges: true,
-            inlay_hints: true,
-            file_rename: true,
-            // Cross-file diagnostic groups are off by default — they scan
-            // every Vue file in the workspace, which is too slow for the
-            // default editor experience. Opt in via
-            // `languageServer.crossFile = true`.
-            cross_file: false,
-        }
-    }
-}
-
-/// Batch type check result cache.
-#[cfg(feature = "native")]
-pub struct BatchTypeCheckCache {
-    /// Diagnostics per file.
-    pub diagnostics: DashMap<PathBuf, Vec<vize_canon::BatchDiagnostic>>,
-    /// Whether the cache is valid.
-    pub valid: std::sync::atomic::AtomicBool,
-}
-
-#[cfg(feature = "native")]
-impl BatchTypeCheckCache {
-    /// Create a new empty cache.
-    pub fn new() -> Self {
-        Self {
-            diagnostics: DashMap::new(),
-            valid: std::sync::atomic::AtomicBool::new(false),
-        }
-    }
-
-    /// Invalidate the cache.
-    pub fn invalidate(&self) {
-        self.valid.store(false, std::sync::atomic::Ordering::SeqCst);
-        self.diagnostics.clear();
-    }
-
-    /// Check if the cache is valid.
-    pub fn is_valid(&self) -> bool {
-        self.valid.load(std::sync::atomic::Ordering::SeqCst)
-    }
-
-    /// Mark the cache as valid.
-    pub fn mark_valid(&self) {
-        self.valid.store(true, std::sync::atomic::Ordering::SeqCst);
-    }
-
-    /// Get diagnostics for a file.
-    pub fn get_diagnostics(&self, path: &PathBuf) -> Vec<vize_canon::BatchDiagnostic> {
-        self.diagnostics
-            .get(path)
-            .map(|d| d.clone())
-            .unwrap_or_default()
-    }
-
-    /// Set diagnostics for a file.
-    pub fn set_diagnostics(&self, path: PathBuf, diagnostics: Vec<vize_canon::BatchDiagnostic>) {
-        self.diagnostics.insert(path, diagnostics);
-    }
-}
-
-#[cfg(feature = "native")]
-impl Default for BatchTypeCheckCache {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Server state containing all runtime data.
-pub struct ServerState {
-    /// Document store for managing open documents
-    pub documents: DocumentStore,
-    /// Virtual code generator (reusable)
-    virtual_gen: RwLock<VirtualCodeGenerator>,
-    /// Cached virtual documents per file
-    virtual_docs_cache: DashMap<Url, VirtualDocuments>,
-    /// Parsed metadata for imported components, keyed by resolved path.
-    /// Lets template completion skip re-reading + re-parsing + re-analyzing an
-    /// imported component on every keystroke; entries are invalidated by the
-    /// component file's length + modification time.
-    component_metadata_cache:
-        DashMap<PathBuf, crate::ide::completion::template::CachedComponentMetadata>,
-    /// Enabled LSP feature surface.
-    lsp_features: RwLock<LspFeatureConfig>,
-    /// Fast path for checking whether type-aware features are enabled.
-    lsp_typecheck_enabled: AtomicBool,
-    /// Type checker options shared by LSP diagnostics.
-    type_checker_config: RwLock<TypeCheckerConfig>,
-    /// Vue 3 Options API binding-resolution opt-in from config.
-    type_checker_options_api: RwLock<bool>,
-    /// Vue 2.7 / Nuxt 2 type checker compatibility flag from config.
-    type_checker_legacy_vue2: RwLock<bool>,
-    /// Linter options shared by LSP diagnostics.
-    linter_config: RwLock<LinterConfig>,
-    /// Formatting options (loaded from vize.config.json)
-    #[cfg(feature = "glyph")]
-    format_options: RwLock<vize_glyph::FormatOptions>,
-    /// Corsa bridge for native TypeScript language features (lazy initialized)
-    #[cfg(feature = "native")]
-    corsa_bridge: OnceLock<Arc<CorsaBridge>>,
-    /// Serializes Corsa bridge initialization without tying us to a runtime.
-    #[cfg(feature = "native")]
-    corsa_init_lock: AsyncMutex<()>,
-    /// Flag to track if Corsa initialization has been attempted and failed
-    #[cfg(feature = "native")]
-    corsa_init_failed: std::sync::atomic::AtomicBool,
-    /// Human-readable reason recorded on Corsa init failure, used by
-    /// `corsa_init_failure()` to surface diagnostic context to handlers and
-    /// tests. Populated alongside `corsa_init_failed` (see #751).
-    #[cfg(feature = "native")]
-    corsa_init_failure_reason: RwLock<Option<Arc<str>>>,
-    /// True once the LSP server has shown the user a one-shot
-    /// `window/showMessage` explaining that type checking is unavailable.
-    /// Prevents the message from firing once per file.
-    #[cfg(feature = "native")]
-    typecheck_unavailable_notified: std::sync::atomic::AtomicBool,
-    /// Workspace root path
-    #[cfg(feature = "native")]
-    workspace_root: RwLock<Option<PathBuf>>,
-    /// Batch type checker (lazy initialized, sync)
-    #[cfg(feature = "native")]
-    batch_checker: OnceLock<Arc<RwLock<BatchTypeChecker>>>,
-    /// Batch type check result cache
-    #[cfg(feature = "native")]
-    batch_cache: BatchTypeCheckCache,
-}
-
-impl Default for ServerState {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl ServerState {
-    /// Create a new server state.
-    pub fn new() -> Self {
-        let default_features = LspFeatureConfig::default();
-        Self {
-            documents: DocumentStore::new(),
-            virtual_gen: RwLock::new(VirtualCodeGenerator::new()),
-            virtual_docs_cache: DashMap::new(),
-            component_metadata_cache: DashMap::new(),
-            lsp_features: RwLock::new(default_features),
-            lsp_typecheck_enabled: AtomicBool::new(default_features.typecheck),
-            type_checker_config: RwLock::new(TypeCheckerConfig::default()),
-            type_checker_options_api: RwLock::new(false),
-            type_checker_legacy_vue2: RwLock::new(false),
-            linter_config: RwLock::new(LinterConfig::default()),
-            #[cfg(feature = "glyph")]
-            format_options: RwLock::new(vize_glyph::FormatOptions::default()),
-            #[cfg(feature = "native")]
-            corsa_bridge: OnceLock::new(),
-            #[cfg(feature = "native")]
-            corsa_init_lock: AsyncMutex::new(()),
-            #[cfg(feature = "native")]
-            corsa_init_failed: std::sync::atomic::AtomicBool::new(false),
-            #[cfg(feature = "native")]
-            corsa_init_failure_reason: RwLock::new(None),
-            #[cfg(feature = "native")]
-            typecheck_unavailable_notified: std::sync::atomic::AtomicBool::new(false),
-            #[cfg(feature = "native")]
-            workspace_root: RwLock::new(None),
-            #[cfg(feature = "native")]
-            batch_checker: OnceLock::new(),
-            #[cfg(feature = "native")]
-            batch_cache: BatchTypeCheckCache::new(),
-        }
-    }
-
-    /// Try to claim the right to fire the "type checking unavailable"
-    /// message. Returns true the first time it is called, false thereafter
-    /// for the lifetime of the server. The caller is responsible for
-    /// actually sending the message via the LSP client.
-    #[cfg(feature = "native")]
-    pub fn claim_typecheck_unavailable_notice(&self) -> bool {
-        !self
-            .typecheck_unavailable_notified
-            .swap(true, std::sync::atomic::Ordering::SeqCst)
-    }
-
-    /// Set the workspace root path.
-    #[cfg(feature = "native")]
-    pub fn set_workspace_root(&self, path: PathBuf) {
-        *self.workspace_root.write() = Some(path);
-        // Invalidate batch cache when workspace changes
-        self.batch_cache.invalidate();
-    }
-
-    /// Check whether LSP type checking is enabled.
-    #[inline]
-    pub fn is_lsp_typecheck_enabled(&self) -> bool {
-        self.lsp_typecheck_enabled.load(Ordering::SeqCst)
-    }
-
-    /// Get the enabled LSP feature set.
-    #[inline]
-    pub(crate) fn lsp_features(&self) -> LspFeatureConfig {
-        *self.lsp_features.read()
-    }
-
-    #[inline]
-    pub(crate) fn legacy_vue2_enabled(&self) -> bool {
-        *self.type_checker_legacy_vue2.read() || self.lsp_features().legacy_vue2
-    }
-
-    /// Resolve Vue 3 Options API template bindings. Implied by legacy mode.
-    #[inline]
-    pub(crate) fn options_api_enabled(&self) -> bool {
-        *self.type_checker_options_api.read()
-            || self.lsp_features().options_api
-            || self.legacy_vue2_enabled()
-    }
-
-    /// Check whether LSP lint diagnostics are enabled.
-    #[inline]
-    pub fn is_lsp_lint_enabled(&self) -> bool {
-        self.lsp_features().lint
-    }
-
-    /// Apply LSP initialization options sent by an editor client.
-    pub fn apply_lsp_initialization_options(&self, options: Option<&serde_json::Value>) {
-        let Some(options) = options else {
-            return;
-        };
-
-        match serde_json::from_value::<LspConfigSection>(options.clone()) {
-            Ok(config) => self.apply_lsp_config(config, "initializationOptions"),
-            Err(error) => {
-                tracing::warn!(
-                    "Failed to parse LSP initializationOptions: {}. Keeping current LSP options.",
-                    error
-                );
-            }
-        }
-    }
-
-    /// Get a clone of the current type checker config.
-    #[inline]
-    pub fn get_type_checker_config(&self) -> TypeCheckerConfig {
-        self.type_checker_config.read().clone()
-    }
-
-    /// Get a clone of the current linter config.
-    #[inline]
-    pub fn get_linter_config(&self) -> LinterConfig {
-        self.linter_config.read().clone()
-    }
-
-    fn apply_type_checker_config(&self, config: TypeCheckerConfig, source: &str) {
-        *self.type_checker_config.write() = config;
-        tracing::info!("Loaded type checker config from {}", source);
-    }
-
-    fn apply_config_features(&self, features: vize_carton::config::ConfigFeatureFlags) {
-        *self.type_checker_options_api.write() = features.type_checker_options_api;
-        *self.type_checker_legacy_vue2.write() = features.type_checker_legacy_vue2;
-        if let Some(enabled) = features.language_server_legacy_vue2 {
-            let mut lsp_features = self.lsp_features.write();
-            lsp_features.legacy_vue2 = enabled;
-        }
-    }
-
-    fn apply_linter_config(&self, config: LinterConfig, source: &str) {
-        *self.linter_config.write() = config;
-        tracing::info!("Loaded linter config from {}", source);
-    }
-
-    fn apply_lsp_config(&self, config: LspConfigSection, source: &str) {
-        let mut features = self.lsp_features.write();
-        config.apply_to(&mut features);
-        self.lsp_typecheck_enabled
-            .store(features.typecheck, Ordering::SeqCst);
-        tracing::info!("Loaded LSP config from {}: {:?}", source, *features);
-    }
-
-    /// Load all workspace-scoped options from `vize.config.pkl` (preferred) or JSON.
-    pub fn load_workspace_config(&self, dir: &Path) {
-        let (loaded, linter_config) =
-            vize_carton::config::load_config_and_linter_with_features_and_source(Some(dir));
-        if let Some(source_path) = loaded.source_path {
-            let source = source_path.display().to_string();
-            let config = loaded.config;
-            #[cfg(feature = "glyph")]
-            {
-                *self.format_options.write() = format_options_from_config(&config.formatter);
-                tracing::info!("Loaded format config from {}", source);
-            }
-            self.apply_linter_config(linter_config, &source);
-            self.apply_type_checker_config(config.type_checker, &source);
-            self.apply_lsp_config(config.language_server.into(), &source);
-            self.apply_config_features(loaded.features);
-        }
-    }
-
-    /// Load LSP options from `vize.config.pkl` (preferred) or `vize.config.json`.
-    pub fn load_lsp_config(&self, dir: &Path) {
-        let (loaded, linter_config) =
-            vize_carton::config::load_config_and_linter_with_features_and_source(Some(dir));
-        if let Some(source_path) = loaded.source_path {
-            let source = source_path.display().to_string();
-            self.apply_linter_config(linter_config, &source);
-            self.apply_type_checker_config(loaded.config.type_checker, &source);
-            self.apply_lsp_config(loaded.config.language_server.into(), &source);
-            self.apply_config_features(loaded.features);
-        }
-    }
-
-    /// Get the workspace root path.
-    #[cfg(feature = "native")]
-    pub fn get_workspace_root(&self) -> Option<PathBuf> {
-        self.workspace_root.read().clone()
-    }
-
-    /// Get or initialize the batch type checker.
-    #[cfg(feature = "native")]
-    pub fn get_batch_checker(&self) -> Option<Arc<RwLock<BatchTypeChecker>>> {
-        if !self.is_lsp_typecheck_enabled() {
-            return None;
-        }
-
-        let workspace_root = self.get_workspace_root()?;
-
-        // Try to get existing value first
-        if let Some(checker) = self.batch_checker.get() {
-            return Some(checker.clone());
-        }
-
-        // Try to initialize
-        let config = self.get_type_checker_config();
-        let corsa_path = config.runtime_path().map(PathBuf::from);
-        let options = BatchTypeCheckerOptions {
-            tsconfig_path: config.tsconfig.as_ref().map(PathBuf::from),
-            virtual_ts_options: vize_canon::virtual_ts::VirtualTsOptions::default(),
-        };
-
-        match BatchTypeChecker::with_options_and_corsa_path(
-            &workspace_root,
-            options,
-            corsa_path.as_deref(),
-        ) {
-            Ok(mut checker) => {
-                if self.options_api_enabled() {
-                    checker.enable_options_api();
-                }
-                if self.legacy_vue2_enabled() {
-                    checker.enable_legacy_vue2();
-                }
-                let arc = Arc::new(RwLock::new(checker));
-                // get_or_init to handle race condition
-                Some(self.batch_checker.get_or_init(|| arc.clone()).clone())
-            }
-            Err(_) => None,
-        }
-    }
-
-    /// Check if batch type checker is available.
-    #[cfg(feature = "native")]
-    pub fn has_batch_checker(&self) -> bool {
-        self.is_lsp_typecheck_enabled() && self.batch_checker.get().is_some()
-    }
-
-    /// Get the batch type check cache.
-    #[cfg(feature = "native")]
-    pub fn get_batch_cache(&self) -> &BatchTypeCheckCache {
-        &self.batch_cache
-    }
-
-    /// Run batch type checking and update the cache.
-    #[cfg(feature = "native")]
-    pub fn run_batch_type_check(&self) -> Option<vize_canon::BatchTypeCheckResult> {
-        if !self.is_lsp_typecheck_enabled() {
-            return None;
-        }
-
-        let checker = self.get_batch_checker()?;
-        let mut checker_guard = checker.write();
-
-        // Scan project if not already scanned
-        if checker_guard.file_count() == 0 && checker_guard.scan_project().is_err() {
-            return None;
-        }
-
-        // Run type check
-        let result = checker_guard.check_project().ok()?;
-
-        // Update cache
-        self.batch_cache.diagnostics.clear();
-        for diag in &result.diagnostics {
-            self.batch_cache
-                .diagnostics
-                .entry(diag.file.clone())
-                .or_default()
-                .push(diag.clone());
-        }
-        self.batch_cache.mark_valid();
-
-        Some(result)
-    }
-
-    /// Invalidate batch type check cache (e.g., when a file changes).
-    #[cfg(feature = "native")]
-    pub fn invalidate_batch_cache(&self) {
-        self.batch_cache.invalidate();
-    }
-
-    /// Get or initialize the Corsa bridge.
-    ///
-    /// Returns `None` if Corsa is not available or failed to initialize.
-    #[cfg(feature = "native")]
-    pub async fn get_corsa_bridge(&self) -> Option<Arc<CorsaBridge>> {
-        if !self.is_lsp_typecheck_enabled() {
-            tracing::info!(
-                "Skipping Corsa bridge initialization because LSP typecheck is disabled"
-            );
-            return None;
-        }
-
-        // If already initialized successfully, return it
-        if let Some(bridge) = self.corsa_bridge.get() {
-            return Some(bridge.clone());
-        }
-
-        // If initialization already failed, don't retry
-        if self.corsa_init_failed.load(Ordering::SeqCst) {
-            return None;
-        }
-
-        let _guard = self.corsa_init_lock.lock().await;
-
-        // Another request may have completed initialization while we were waiting.
-        if let Some(bridge) = self.corsa_bridge.get() {
-            return Some(bridge.clone());
-        }
-
-        if self.corsa_init_failed.load(Ordering::SeqCst) {
-            return None;
-        }
-
-        // Get workspace root for Corsa configuration.
-        let workspace_root = self.get_workspace_root();
-        let type_checker_config = self.get_type_checker_config();
-
-        let config = CorsaBridgeConfig {
-            corsa_path: type_checker_config.runtime_path().map(PathBuf::from),
-            working_dir: workspace_root,
-            timeout_ms: 30000, // Corsa needs time to build project state on first load.
-            ..Default::default()
-        };
-        let working_dir = config.working_dir.clone();
-        let corsa_path = config.corsa_path.clone();
-        let bridge = CorsaBridge::with_config(config);
-
-        match crate::runtime::timeout(std::time::Duration::from_secs(5), bridge.spawn()).await {
-            Ok(Ok(())) => {
-                tracing::info!("corsa bridge initialized successfully");
-                let bridge = Arc::new(bridge);
-                let _ = self.corsa_bridge.set(bridge.clone());
-                Some(bridge)
-            }
-            Ok(Err(e)) => {
-                let reason = vize_carton::cstr!(
-                    "spawn failed: {e} (working_dir={working_dir:?}, corsa_path={corsa_path:?})"
-                );
-                tracing::warn!("corsa bridge {}", reason);
-                self.record_corsa_init_failure(reason.as_str());
-                None
-            }
-            Err(_) => {
-                let reason = vize_carton::cstr!(
-                    "spawn timed out after 5s (working_dir={working_dir:?}, corsa_path={corsa_path:?})"
-                );
-                tracing::warn!("corsa bridge {}", reason);
-                self.record_corsa_init_failure(reason.as_str());
-                None
-            }
-        }
-    }
-
-    /// Read the human-readable reason recorded the last time Corsa bridge
-    /// initialization failed. Returns `None` if init has not failed.
-    ///
-    /// Used by handlers and tests to diagnose why the editor session fell
-    /// back to the heuristic completion path (see #751).
-    #[cfg(feature = "native")]
-    pub fn corsa_init_failure(&self) -> Option<Arc<str>> {
-        self.corsa_init_failure_reason.read().clone()
-    }
-
-    #[cfg(feature = "native")]
-    fn record_corsa_init_failure(&self, reason: &str) {
-        *self.corsa_init_failure_reason.write() = Some(Arc::from(reason));
-        self.corsa_init_failed.store(true, Ordering::SeqCst);
-    }
-
-    /// Check if the Corsa bridge is available (without initializing).
-    #[cfg(feature = "native")]
-    pub fn has_corsa_bridge(&self) -> bool {
-        self.is_lsp_typecheck_enabled() && self.corsa_bridge.get().is_some()
-    }
-
-    /// Generate and cache virtual documents for a document.
-    pub fn update_virtual_docs(&self, uri: &Url, content: &str) {
-        if uri.path().ends_with(".art.vue") {
-            self.update_art_virtual_docs(uri, content);
-            return;
-        }
-
-        if is_standalone_html_path(uri.path()) {
-            self.update_standalone_html_virtual_docs(uri, content);
-            return;
-        }
-
-        let options = vize_atelier_sfc::SfcParseOptions {
-            filename: uri.path().to_string().into(),
-            ..Default::default()
-        };
-
-        let Ok(descriptor) = vize_atelier_sfc::parse_sfc(content, options) else {
-            self.remove_virtual_docs(uri);
-            return;
-        };
-
-        let base_uri = uri.path();
-        let virtual_docs = self.virtual_gen.write().generate(&descriptor, base_uri);
-        self.virtual_docs_cache.insert(uri.clone(), virtual_docs);
-    }
-
-    /// Generate and cache virtual documents for standalone HTML files.
-    fn update_standalone_html_virtual_docs(&self, uri: &Url, content: &str) {
-        use crate::virtual_code::{TemplateCodeGenerator, VirtualDocuments};
-
-        let allocator = vize_carton::Bump::new();
-        let (ast, _errors) = vize_armature::parse(&allocator, content);
-        let base_uri = uri.path();
-
-        let mut template_gen = TemplateCodeGenerator::new();
-        template_gen.set_block_offset(0);
-        let mut template_doc = template_gen.generate(&ast, content);
-        template_doc.uri = vize_carton::cstr!("{base_uri}.__template.ts").to_string();
-
-        let mut docs = VirtualDocuments::new();
-        docs.template = Some(template_doc);
-        self.virtual_docs_cache.insert(uri.clone(), docs);
-    }
-
-    /// Generate and cache virtual documents for an art file (*.art.vue).
-    ///
-    /// Uses the default variant's template as the synthetic template block,
-    /// and generates virtual docs for script_setup if present.
-    fn update_art_virtual_docs(&self, uri: &Url, content: &str) {
-        use crate::virtual_code::{ScriptCodeGenerator, TemplateCodeGenerator, VirtualDocuments};
-
-        let allocator = vize_carton::Bump::new();
-        let Ok(art_desc) =
-            vize_musea::parse_art(&allocator, content, vize_musea::ArtParseOptions::default())
-        else {
-            self.remove_virtual_docs(uri);
-            return;
-        };
-
-        let base_uri = uri.path();
-        let mut docs = VirtualDocuments::new();
-
-        // Generate one virtual template per variant so editor features remain correct even when
-        // the cursor is inside a non-default variant.
-        docs.art_templates.resize(art_desc.variants.len(), None);
-
-        for (index, variant) in art_desc.variants.iter().enumerate() {
-            let template_content = variant.template;
-            if template_content.trim().is_empty() {
-                continue;
-            }
-
-            let template_allocator = vize_carton::Bump::new();
-            let (ast, _errors) = vize_armature::parse(&template_allocator, template_content);
-
-            let template_ptr = template_content.as_ptr() as usize;
-            let source_ptr = content.as_ptr() as usize;
-            let block_offset = (template_ptr - source_ptr) as u32;
-
-            let mut template_gen = TemplateCodeGenerator::new();
-            template_gen.set_block_offset(block_offset);
-            let mut template_doc = template_gen.generate(&ast, template_content);
-            template_doc.uri =
-                vize_carton::cstr!("{base_uri}.art_variant_{index}.template.ts").to_string();
-
-            if variant.is_default || docs.template.is_none() {
-                docs.template = Some(template_doc.clone());
-            }
-
-            docs.art_templates[index] = Some(template_doc);
-        }
-
-        // Generate script_setup virtual doc using SFC parser
-        // (SFC parser handles script blocks even in art files)
-        let sfc_options = vize_atelier_sfc::SfcParseOptions {
-            filename: uri.path().to_string().into(),
-            ..Default::default()
-        };
-        if let Ok(descriptor) = vize_atelier_sfc::parse_sfc(content, sfc_options) {
-            if let Some(ref script_setup) = descriptor.script_setup {
-                let isolate = art_script_setup_isolated(script_setup);
-                let mut script_doc = generate_art_script_setup_virtual_doc(
-                    base_uri,
-                    script_setup.content.as_ref(),
-                    script_setup.loc.start,
-                    art_desc.variants.len(),
-                    isolate,
-                );
-                script_doc.uri = vize_carton::cstr!("{base_uri}.__script_setup.ts").to_string();
-                docs.script_setup = Some(script_doc);
-            }
-            if let Some(ref script) = descriptor.script {
-                let mut script_gen = ScriptCodeGenerator::new();
-                let mut script_doc = script_gen.generate(script, false);
-                script_doc.uri = vize_carton::cstr!("{base_uri}.__script.ts").to_string();
-                docs.script = Some(script_doc);
-            }
-        }
-
-        self.virtual_docs_cache.insert(uri.clone(), docs);
-    }
-
-    /// Get cached virtual documents for a document.
-    pub fn get_virtual_docs(
-        &self,
-        uri: &Url,
-    ) -> Option<dashmap::mapref::one::Ref<'_, Url, VirtualDocuments>> {
-        self.virtual_docs_cache.get(uri)
-    }
-
-    /// Remove cached virtual documents when a document is closed.
-    pub fn remove_virtual_docs(&self, uri: &Url) {
-        self.virtual_docs_cache.remove(uri);
-    }
-
-    /// Clear all cached virtual documents.
-    pub fn clear_virtual_docs(&self) {
-        self.virtual_docs_cache.clear();
-    }
-
-    /// Cache of parsed imported-component metadata, keyed by resolved path.
-    /// Used by template completion to avoid re-parsing imported components on
-    /// every keystroke. Callers handle staleness via the entry's file stamp.
-    pub(crate) fn component_metadata_cache(
-        &self,
-    ) -> &DashMap<PathBuf, crate::ide::completion::template::CachedComponentMetadata> {
-        &self.component_metadata_cache
-    }
-
-    /// Get a clone of the current format options.
-    #[cfg(feature = "glyph")]
-    #[inline]
-    pub fn get_format_options(&self) -> vize_glyph::FormatOptions {
-        self.format_options.read().clone()
-    }
-
-    /// Load format options from `vize.config.json` in the given directory.
-    #[cfg(feature = "glyph")]
-    pub fn load_format_config(&self, dir: &Path) {
-        let loaded = vize_carton::config::load_config_with_source(Some(dir));
-        if let Some(source_path) = loaded.source_path {
-            *self.format_options.write() = format_options_from_config(&loaded.config.formatter);
-            tracing::info!("Loaded format config from {}", source_path.display());
-        }
-    }
-}
-
-fn art_script_setup_isolated(script_setup: &vize_atelier_sfc::SfcScriptBlock<'_>) -> bool {
-    !script_setup
-        .attrs
-        .get("isolate")
-        .is_some_and(|value| value.as_ref().eq_ignore_ascii_case("false"))
-}
-
-fn generate_art_script_setup_virtual_doc(
-    base_uri: &str,
-    script: &str,
-    source_start: usize,
-    variant_count: usize,
-    isolate: bool,
-) -> crate::virtual_code::VirtualDocument {
-    use crate::virtual_code::{
-        MappingFeatures, SourceMap, SourceMapping, SourceRange, VirtualDocument, VirtualLanguage,
-        analyze_art_script_setup,
-    };
-
-    let parts = analyze_art_script_setup(script, source_start, isolate);
-    let mut content = String::new();
-    let mut mappings = Vec::new();
-
-    content.push_str("// Virtual TypeScript for .art.vue <script setup>\n");
-    content.push_str("// Generated by vize_maestro\n\n");
-
-    for import in &parts.shared_imports {
-        let generated_start = content.len() as u32;
-        content.push_str(&import.text);
-        let generated_end = content.len() as u32;
-        mappings.push(SourceMapping::with_features(
-            SourceRange::new(import.source_start as u32, import.source_end as u32),
-            SourceRange::new(generated_start, generated_end),
-            MappingFeatures::all(),
-        ));
-        if !content.ends_with('\n') {
-            content.push('\n');
-        }
-    }
-    if !content.ends_with("\n\n") {
-        content.push('\n');
-    }
-
-    if parts.isolate {
-        let count = variant_count.max(1);
-        for index in 0..count {
-            vize_carton::append!(content, "function __VIZE_art_variant_{index}_setup() {{\n");
-            for chunk in &parts.isolated_body {
-                let generated_start = content.len() as u32;
-                content.push_str(&chunk.text);
-                let generated_end = content.len() as u32;
-                mappings.push(SourceMapping::with_features(
-                    SourceRange::new(chunk.source_start as u32, chunk.source_end as u32),
-                    SourceRange::new(generated_start, generated_end),
-                    MappingFeatures::all(),
-                ));
-                if !content.ends_with('\n') {
-                    content.push('\n');
-                }
-            }
-            content.push_str("}\n\n");
-        }
-    } else {
-        content.push_str("// Shared <script setup isolate=\"false\">\n");
-        for chunk in &parts.isolated_body {
-            let generated_start = content.len() as u32;
-            content.push_str(&chunk.text);
-            let generated_end = content.len() as u32;
-            mappings.push(SourceMapping::with_features(
-                SourceRange::new(chunk.source_start as u32, chunk.source_end as u32),
-                SourceRange::new(generated_start, generated_end),
-                MappingFeatures::all(),
-            ));
-            if !content.ends_with('\n') {
-                content.push('\n');
-            }
-        }
-    }
-
-    VirtualDocument {
-        uri: vize_carton::cstr!("{base_uri}.__script_setup.ts").to_string(),
-        content,
-        language: VirtualLanguage::ScriptSetup,
-        source_map: SourceMap::from_mappings(mappings),
-    }
-}
-
-#[cfg(feature = "glyph")]
-fn format_options_from_config(config: &FormatterConfig) -> vize_glyph::FormatOptions {
-    vize_glyph::FormatOptions {
-        print_width: config.print_width,
-        tab_width: config.tab_width,
-        use_tabs: config.use_tabs,
-        semi: config.semi,
-        single_quote: config.single_quote,
-        jsx_single_quote: config.jsx_single_quote,
-        trailing_comma: match config.trailing_comma {
-            vize_carton::config::TrailingComma::None => vize_glyph::TrailingComma::None,
-            vize_carton::config::TrailingComma::Es5 => vize_glyph::TrailingComma::Es5,
-            vize_carton::config::TrailingComma::All => vize_glyph::TrailingComma::All,
-        },
-        bracket_spacing: config.bracket_spacing,
-        bracket_same_line: config.bracket_same_line,
-        arrow_parens: match config.arrow_parens {
-            vize_carton::config::ArrowParens::Always => vize_glyph::ArrowParens::Always,
-            vize_carton::config::ArrowParens::Avoid => vize_glyph::ArrowParens::Avoid,
-        },
-        end_of_line: match config.end_of_line {
-            vize_carton::config::EndOfLine::Lf => vize_glyph::EndOfLine::Lf,
-            vize_carton::config::EndOfLine::Crlf => vize_glyph::EndOfLine::Crlf,
-            vize_carton::config::EndOfLine::Cr => vize_glyph::EndOfLine::Cr,
-            vize_carton::config::EndOfLine::Auto => vize_glyph::EndOfLine::Auto,
-        },
-        quote_props: match config.quote_props {
-            vize_carton::config::QuoteProps::AsNeeded => vize_glyph::QuoteProps::AsNeeded,
-            vize_carton::config::QuoteProps::Consistent => vize_glyph::QuoteProps::Consistent,
-            vize_carton::config::QuoteProps::Preserve => vize_glyph::QuoteProps::Preserve,
-        },
-        single_attribute_per_line: config.single_attribute_per_line,
-        vue_indent_script_and_style: config.vue_indent_script_and_style,
-        sort_attributes: config.sort_attributes,
-        attribute_sort_order: match config.attribute_sort_order {
-            vize_carton::config::AttributeSortOrder::Alphabetical => {
-                vize_glyph::AttributeSortOrder::Alphabetical
-            }
-            vize_carton::config::AttributeSortOrder::AsWritten => {
-                vize_glyph::AttributeSortOrder::AsWritten
-            }
-        },
-        merge_bind_and_non_bind_attrs: config.merge_bind_and_non_bind_attrs,
-        max_attributes_per_line: config.max_attributes_per_line,
-        attribute_groups: config.attribute_groups.clone(),
-        normalize_directive_shorthands: config.normalize_directive_shorthands,
-        sort_blocks: config.sort_blocks,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::ServerState;
-    use crate::virtual_code::{ArtCursorPosition, BlockType, find_art_block_at_offset};
-    use tower_lsp::lsp_types::Url;
-
-    #[cfg(feature = "native")]
-    #[test]
-    fn corsa_init_failure_is_recorded() {
-        let state = ServerState::new();
-        assert!(state.corsa_init_failure().is_none());
-        state.record_corsa_init_failure("spawn failed: missing tsgo");
-        let reason = state
-            .corsa_init_failure()
-            .expect("failure reason should be recorded");
-        assert!(
-            reason.contains("spawn failed"),
-            "reason should preserve the recorded message: {reason}"
-        );
-    }
-
-    #[test]
-    fn default_format_options() {
-        let state = ServerState::new();
-        let opts = state.get_format_options();
-        assert_eq!(opts.print_width, 100);
-        assert_eq!(opts.tab_width, 2);
-        assert!(!opts.use_tabs);
-        assert!(opts.semi);
-        assert!(!opts.single_quote);
-        assert!(opts.sort_attributes);
-        assert!(opts.normalize_directive_shorthands);
-    }
-
-    #[test]
-    fn load_format_config_no_file() {
-        let dir = tempfile::tempdir().unwrap();
-        let state = ServerState::new();
-        state.load_format_config(dir.path());
-        // options remain default
-        let opts = state.get_format_options();
-        assert_eq!(opts.print_width, 100);
-    }
-
-    #[test]
-    fn load_format_config_from_file() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("vize.config.json"),
-            r#"{
-                "fmt": {
-                    "printWidth": 80,
-                    "tabWidth": 4,
-                    "useTabs": true,
-                    "semi": false,
-                    "singleQuote": true
-                }
-            }"#,
-        )
-        .unwrap();
-
-        let state = ServerState::new();
-        state.load_format_config(dir.path());
-        let opts = state.get_format_options();
-        assert_eq!(opts.print_width, 80);
-        assert_eq!(opts.tab_width, 4);
-        assert!(opts.use_tabs);
-        assert!(!opts.semi);
-        assert!(opts.single_quote);
-    }
-
-    #[test]
-    fn load_format_config_partial() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("vize.config.json"),
-            r#"{ "fmt": { "printWidth": 120 } }"#,
-        )
-        .unwrap();
-
-        let state = ServerState::new();
-        state.load_format_config(dir.path());
-        let opts = state.get_format_options();
-        assert_eq!(opts.print_width, 120);
-        // defaults preserved
-        assert_eq!(opts.tab_width, 2);
-        assert!(opts.semi);
-    }
-
-    #[test]
-    fn load_format_config_no_fmt_section() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("vize.config.json"),
-            r#"{ "check": { "globals": ["$t"] } }"#,
-        )
-        .unwrap();
-
-        let state = ServerState::new();
-        state.load_format_config(dir.path());
-        // no fmt section → options remain default
-        let opts = state.get_format_options();
-        assert_eq!(opts.print_width, 100);
-    }
-
-    #[test]
-    fn load_format_config_invalid_json() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(dir.path().join("vize.config.json"), "not valid json").unwrap();
-
-        let state = ServerState::new();
-        state.load_format_config(dir.path());
-        // options remain default
-        let opts = state.get_format_options();
-        assert_eq!(opts.print_width, 100);
-    }
-
-    #[test]
-    fn lsp_features_enable_non_opinionated_defaults() {
-        let state = ServerState::new();
-        let features = state.lsp_features();
-        assert!(features.lint);
-        assert!(features.typecheck);
-        assert!(features.ecosystem);
-        assert!(features.completion);
-        assert!(features.code_actions);
-        assert!(!features.formatting);
-        assert!(state.is_lsp_lint_enabled());
-        assert!(state.is_lsp_typecheck_enabled());
-    }
-
-    #[test]
-    fn load_lsp_config_from_json() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("vize.config.json"),
-            r#"{
-                "lsp": {
-                    "lint": true,
-                    "typecheck": true,
-                    "editor": true,
-                    "formatting": false
-                }
-            }"#,
-        )
-        .unwrap();
-
-        let state = ServerState::new();
-        state.load_lsp_config(dir.path());
-        let features = state.lsp_features();
-        assert!(features.lint);
-        assert!(features.typecheck);
-        assert!(features.ecosystem);
-        assert!(features.completion);
-        assert!(features.definition);
-        assert!(!features.formatting);
-    }
-
-    #[test]
-    fn load_lsp_config_updates_type_checker_config() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("vize.config.json"),
-            r#"{
-                "typeChecker": {
-                    "strict": true,
-                    "checkProps": false,
-                    "checkEmits": false,
-                    "tsconfig": "tsconfig.app.json",
-                    "corsaPath": "./node_modules/.bin/corsa"
-                }
-            }"#,
-        )
-        .unwrap();
-
-        let state = ServerState::new();
-        state.load_lsp_config(dir.path());
-        let config = state.get_type_checker_config();
-        assert!(config.strict);
-        assert!(!config.check_props);
-        assert!(!config.check_emits);
-        assert_eq!(config.tsconfig.as_deref(), Some("tsconfig.app.json"));
-        assert_eq!(config.runtime_path(), Some("./node_modules/.bin/corsa"));
-    }
-
-    #[test]
-    fn options_api_disabled_by_default() {
-        let state = ServerState::new();
-        assert!(
-            !state.options_api_enabled(),
-            "Options API resolution must be opt-in: zero cost on the default Vue 3 path"
-        );
-    }
-
-    #[test]
-    fn type_checker_options_api_opt_in_from_config() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("vize.config.json"),
-            r#"{ "typeChecker": { "optionsApi": true } }"#,
-        )
-        .unwrap();
-
-        let state = ServerState::new();
-        state.load_workspace_config(dir.path());
-        assert!(
-            state.options_api_enabled(),
-            "typeChecker.optionsApi should enable Options API binding resolution in the LSP"
-        );
-    }
-
-    #[test]
-    fn legacy_vue2_config_implies_options_api() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("vize.config.json"),
-            r#"{ "typeChecker": { "legacyVue2": true } }"#,
-        )
-        .unwrap();
-
-        let state = ServerState::new();
-        state.load_workspace_config(dir.path());
-        assert!(
-            state.legacy_vue2_enabled(),
-            "typeChecker.legacyVue2 should enable Vue 2 compatibility"
-        );
-        assert!(
-            state.options_api_enabled(),
-            "legacy Vue 2 mode is a superset of Options API binding resolution"
-        );
-    }
-
-    #[test]
-    fn load_lsp_config_updates_linter_config() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("vize.config.json"),
-            r#"{
-                "linter": {
-                    "preset": "opinionated",
-                    "rules": {
-                        "vue/prop-name-casing": "off"
-                    }
-                }
-            }"#,
-        )
-        .unwrap();
-
-        let state = ServerState::new();
-        state.load_lsp_config(dir.path());
-        let config = state.get_linter_config();
-        assert_eq!(config.preset.as_deref(), Some("opinionated"));
-        assert_eq!(config.disabled_rules(), ["vue/prop-name-casing"]);
-    }
-
-    #[test]
-    fn load_lsp_config_invalid_json_keeps_default() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(dir.path().join("vize.config.json"), "not valid json").unwrap();
-
-        let state = ServerState::new();
-        state.load_lsp_config(dir.path());
-        assert_eq!(state.lsp_features(), super::LspFeatureConfig::default());
-    }
-
-    #[test]
-    fn apply_lsp_initialization_options() {
-        let state = ServerState::new();
-        let options = serde_json::json!({
-            "lint": true,
-            "codeActions": true,
-            "definition": true,
-            "ecosystem": true
-        });
-
-        state.apply_lsp_initialization_options(Some(&options));
-
-        let features = state.lsp_features();
-        assert!(features.lint);
-        assert!(features.code_actions);
-        assert!(features.definition);
-        assert!(features.ecosystem);
-        assert!(features.typecheck);
-    }
-
-    #[test]
-    fn update_art_virtual_docs_tracks_non_default_variants_separately() {
-        let state = ServerState::new();
-        let uri = Url::parse("file:///Button.art.vue").unwrap();
-        let source = r#"<script setup lang="ts">
-const primaryLabel = ref('primary')
-const secondaryLabel = ref('secondary')
-</script>
-
-<art title="Button" component="./Button.vue">
-  <variant name="Primary" default>
-    <Button :label="primaryLabel" />
-  </variant>
-  <variant name="Secondary">
-    <Button :label="secondaryLabel" />
-  </variant>
-</art>
-"#;
-
-        state
-            .documents
-            .open(uri.clone(), source.to_string(), 1, "art-vue".to_string());
-        state.update_virtual_docs(&uri, source);
-
-        let virtual_docs = state.get_virtual_docs(&uri).unwrap();
-        assert_eq!(virtual_docs.art_templates.len(), 2);
-
-        let default_template = virtual_docs.template.as_ref().unwrap();
-        let secondary_template = virtual_docs.art_template(1).unwrap();
-
-        assert!(default_template.content.contains("primaryLabel"));
-        assert!(secondary_template.content.contains("secondaryLabel"));
-        assert!(!secondary_template.uri.ends_with(".__template.ts"));
-        assert!(
-            secondary_template
-                .uri
-                .contains(".art_variant_1.template.ts")
-        );
-
-        let offset = source.rfind("secondaryLabel").unwrap() + 1;
-        let info = match find_art_block_at_offset(source, offset) {
-            Some(BlockType::Art(ArtCursorPosition::VariantTemplate(info))) => info,
-            other => panic!("expected secondary variant template, got {other:?}"),
-        };
-
-        let generated_offset = secondary_template
-            .source_map
-            .to_generated(info.relative_offset as u32);
-        assert!(generated_offset.is_some());
-    }
-
-    #[test]
-    fn update_art_virtual_docs_isolates_script_setup_per_variant() {
-        let state = ServerState::new();
-        let uri = Url::parse("file:///Counter.art.vue").unwrap();
-        let source = r#"<script setup lang="ts">
-import { computed, ref } from "vue";
-defineArt("./Counter.vue", { title: "Counter" });
-const count = ref(0)
-const doubled = computed(() => count.value * 2)
-</script>
-
-<art>
-  <variant name="First">
-    <Counter :count="doubled" />
-  </variant>
-  <variant name="Second">
-    <Counter :count="count" />
-  </variant>
-</art>
-"#;
-
-        state
-            .documents
-            .open(uri.clone(), source.to_string(), 1, "art-vue".to_string());
-        state.update_virtual_docs(&uri, source);
-
-        let virtual_docs = state.get_virtual_docs(&uri).unwrap();
-        let script_setup = virtual_docs.script_setup.as_ref().unwrap();
-        assert!(
-            script_setup
-                .content
-                .contains("function __VIZE_art_variant_0_setup()")
-        );
-        assert!(
-            script_setup
-                .content
-                .contains("function __VIZE_art_variant_1_setup()")
-        );
-        assert!(!script_setup.content.contains("defineArt"));
-
-        let state_offset = source.find("doubled = computed").unwrap();
-        assert!(
-            script_setup
-                .source_map
-                .to_generated(state_offset as u32)
-                .is_some()
-        );
-    }
-
-    #[test]
-    fn update_art_virtual_docs_keeps_script_setup_shared_when_isolate_false() {
-        let state = ServerState::new();
-        let uri = Url::parse("file:///Counter.art.vue").unwrap();
-        let source = r#"<script setup lang="ts" isolate="false">
-const count = ref(0)
-</script>
-
-<art title="Counter">
-  <variant name="First">
-    <Counter :count="count" />
-  </variant>
-  <variant name="Second">
-    <Counter :count="count" />
-  </variant>
-</art>
-"#;
-
-        state
-            .documents
-            .open(uri.clone(), source.to_string(), 1, "art-vue".to_string());
-        state.update_virtual_docs(&uri, source);
-
-        let virtual_docs = state.get_virtual_docs(&uri).unwrap();
-        let script_setup = virtual_docs.script_setup.as_ref().unwrap();
-
-        assert!(script_setup.content.contains("isolate=\"false\""));
-        assert!(!script_setup.content.contains("__VIZE_art_variant_0_setup"));
-        assert!(!script_setup.content.contains("__VIZE_art_variant_1_setup"));
-    }
-
-    #[test]
-    fn update_virtual_docs_generates_standalone_html_template_doc() {
-        let state = ServerState::new();
-        let uri = Url::parse("file:///index.html").unwrap();
-        let source = r#"<div v-scope="{ count: 0 }">{{ count }}</div>"#;
-
-        state.update_virtual_docs(&uri, source);
-
-        let virtual_docs = state.get_virtual_docs(&uri).unwrap();
-        let template = virtual_docs.template.as_ref().unwrap();
-        assert!(template.uri.ends_with("index.html.__template.ts"));
-        assert!(template.content.contains("count"));
-    }
-
-    #[test]
-    fn update_virtual_docs_removes_cache_after_sfc_parse_failure() {
-        let state = ServerState::new();
-        let uri = Url::parse("file:///Broken.vue").unwrap();
-        let valid_source = r#"<script setup lang="ts">
-const message = 'ok'
-</script>
-
-<template>
-  <div>{{ message }}</div>
-</template>
-"#;
-
-        state.update_virtual_docs(&uri, valid_source);
-        assert!(state.get_virtual_docs(&uri).is_some());
-
-        state.update_virtual_docs(&uri, "<template><div></div>");
-        assert!(state.get_virtual_docs(&uri).is_none());
-    }
-
-    #[test]
-    fn update_virtual_docs_removes_cache_after_art_parse_failure() {
-        let state = ServerState::new();
-        let uri = Url::parse("file:///Broken.art.vue").unwrap();
-        let valid_source = r#"<art title="Button" component="./Button.vue">
-  <variant name="Primary" default>
-    <Button />
-  </variant>
-</art>
-"#;
-
-        state.update_virtual_docs(&uri, valid_source);
-        assert!(state.get_virtual_docs(&uri).is_some());
-
-        state.update_virtual_docs(&uri, "<template><div>not an art file</div></template>");
-        assert!(state.get_virtual_docs(&uri).is_none());
-    }
-
-    #[test]
-    #[ignore = "requires pkl runtime installed"]
-    fn load_lsp_config_from_pkl() {
-        let dir = tempfile::tempdir().unwrap();
-        std::fs::write(
-            dir.path().join("vize.config.pkl"),
-            "lsp {\n    lint = true\n    typecheck = true\n}\n",
-        )
-        .unwrap();
-
-        let state = ServerState::new();
-        state.load_lsp_config(dir.path());
-        assert!(state.is_lsp_lint_enabled());
-        assert!(state.is_lsp_typecheck_enabled());
-    }
-}
